test: update conftest.py
